refactor: (#1056) notification stub 위치 수정
